# docs/openapi/evaluator.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Oscanner Evaluator API",
    "summary": "Contributor capability evaluation service",
    "description": "Extract GitHub and Gitee activity, discover contributor identities, evaluate contributors with plugin-defined rubrics, analyze trajectories, and proxy repository runner workflows.",
    "license": {
      "name": "Apache-2.0"
    },
    "version": "0.1.6"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local evaluator service"
    }
  ],
  "paths": {
...
    "/api/runner/run-all": {
      "post": {
        "tags": [
          "runner"
        ],
        "summary": "Run All Steps",
        "description": "Proxy the full clone → explore → run-tests pipeline to the runner service.\n\nForwards the SSE stream from the runner's /api/runner/run-all directly to\nthe caller so progress events and the final status event are all visible.",
        "operationId": "run_all_steps_api_runner_run_all_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunAllRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runner/run-all_poll": {
      "post": {
        "tags": [
          "runner"
        ],
        "summary": "Start Run All Poll",
        "description": "Start a durable clone → explore → test runner job and return its poll URL.",
        "operationId": "start_run_all_poll_api_runner_run_all_poll_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunAllRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runner/run-all_poll/{job_id}": {
      "get": {
        "tags": [
          "runner"
        ],
        "summary": "Get Run All Poll",
        "description": "Return stored events for a durable runner job.",
        "operationId": "get_run_all_poll_api_runner_run_all_poll__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runner/{path}": {
      "get": {
        "tags": [
          "runner"
        ],
        "summary": "Proxy runner GET request",
        "description": "Proxy all /api/runner/* requests to the repos_runner service\n\nArgs:\n    path: The path after /api/runner/\n    request: The incoming request\n\nReturns:\n    Proxied response from repos_runner service",
        "operationId": "proxy_runner_get_api_runner_path",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "runner"
        ],
        "summary": "Proxy runner POST request",
        "description": "Proxy all /api/runner/* requests to the repos_runner service\n\nArgs:\n    path: The path after /api/runner/\n    request: The incoming request\n\nReturns:\n    Proxied response from repos_runner service",
        "operationId": "proxy_runner_post_api_runner_path",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "runner"
        ],
        "summary": "Proxy runner PUT request",
        "description": "Proxy all /api/runner/* requests to the repos_runner service\n\nArgs:\n    path: The path after /api/runner/\n    request: The incoming request\n\nReturns:\n    Proxied response from repos_runner service",
        "operationId": "proxy_runner_put_api_runner_path",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "runner"
        ],
        "summary": "Proxy runner PATCH request",
        "description": "Proxy all /api/runner/* requests to the repos_runner service\n\nArgs:\n    path: The path after /api/runner/\n    request: The incoming request\n\nReturns:\n    Proxied response from repos_runner service",
        "operationId": "proxy_runner_patch_api_runner_path",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "runner"
        ],
        "summary": "Proxy runner DELETE request",
        "description": "Proxy all /api/runner/* requests to the repos_runner service\n\nArgs:\n    path: The path after /api/runner/\n    request: The incoming request\n\nReturns:\n    Proxied response from repos_runner service",
        "operationId": "proxy_runner_delete_api_runner_path",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/checkers/run": {
      "post": {
        "tags": [
          "checkers"
        ],
        "summary": "Run Checker",
        "description": "Run a checker on a specific commit.\n\nArgs:\n    request: CheckerRunRequest with checker_id, platform, owner, repo, commit_sha, and optional files\n    \nReturns:\n    Checker execution result (dict with success, score, passed, total, details, message, etc.)",
        "operationId": "run_checker_api_checkers_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckerRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Run Checker Api Checkers Run Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CheckerRunRequest": {
        "properties": {
          "checker_id": {
            "type": "string",
            "title": "Checker Id"
          },
          "platform": {
            "type": "string",
            "title": "Platform"
          },
          "owner": {
            "type": "string",
            "title": "Owner"
          },
          "repo": {
            "type": "string",
            "title": "Repo"
          },
          "commit_sha": {
            "type": "string",
            "title": "Commit Sha"
          },
          "files": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Files"
          },
          "worktree_base": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Worktree Base"
          }
        },
        "type": "object",
        "required": [
          "checker_id",
          "platform",
          "owner",
          "repo",
          "commit_sha"
        ],
        "title": "CheckerRunRequest",
        "description": "Request model for running a checker."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "RunAllRequest": {
        "properties": {
          "repo_url": {
            "type": "string",
            "title": "Repo Url"
          },
          "sha": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sha"
          },
          "tag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag"
          },
          "branch": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Branch"
          },
          "tag_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag Message"
          },
          "grading_rubric": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grading Rubric",
            "default": "评分规则\n\n你正在评估 Vibe Coding 时代的工程师。请区分“AI搬运工”和“系统构建者”。使用 L1-L5 行为画像作为评分指引：\n\nL1：盲目复制/粘贴，无法解释代码，常见低级错误，没有质量门禁。\nL2：能交付 happy-path，具备基础工程规范、基础测试/lint，但系统性思考较浅。\nL3：能作为单人全栈 MVP 构建者，能重构 AI 生成代码，具备更强类型纪律，关注边界情况。\nL4：团队骨干，能引入质量门禁、防御性校验、CI、文档，并考虑成本和运维。\nL5：领导者/维护者，能定义模式和标准，影响生态，做出深层架构决策。\n\n证据查找方向（优先使用仓库产物，而不是口头声明）：\n\n规格/质量：重构、模块化、输入校验、测试（单元/集成/性质测试）、lint/format、CI。\n可复现性：依赖锁文件、一键运行、Docker/Compose/devcontainer。\n云原生：容器化、IaC、部署配置、资源限制、自动化。\nAI 工程：agent/tooling、结构化 prompt、工具抽象、trace/log、评测数据集。\n专业性：文档/ADR、有意义的 commit/PR、审慎权衡、安全/性能考虑。\n\n评分映射：对每个维度，将观察到的证据映射到粗略 L1-L5，并转换为 0-100 分（L1≈10-30，L2≈30-50，L3≈50-70，L4≈70-85，L5≈85-100）。缺少证据时要保守评分。\n\n轨迹评估上下文：\n\n本评估属于成长轨迹追踪系统。\ncommit 会被分组为评估节点（每个节点至少 10 个 commit，来自 2 周周期）。\n随着工程师学习和进步，分数总体应呈上升趋势。\n只有在存在明确负面证据时才降低分数（质量回退、引入 bug、反模式等）。\n如果提供了上一检查点分数，请将其作为对比基线。\n如果当前工作保持了与上一阶段相近的质量，分数应持平或略有提升。\n显著提分需要明确证据证明出现了新能力或实践改进。"
          },
          "skip_clone": {
            "type": "boolean",
            "title": "Skip Clone",
            "default": false
          },
          "skip_explore": {
            "type": "boolean",
            "title": "Skip Explore",
            "default": false
          },
          "clone_timeout": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Clone Timeout",
            "default": 300
          },
          "setup_timeout": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Setup Timeout",
            "default": 300
          },
          "test_timeout": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Test Timeout",
            "default": 600
          },
          "pipeline_timeout": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Pipeline Timeout",
            "default": 1800
          }
        },
        "type": "object",
        "required": [
          "repo_url"
        ],
        "title": "RunAllRequest",
        "description": "Request model for the combined clone → explore → test pipeline"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}